:recycle: Deprecate CalculatorViewModel
